requirements.txt
beautifulsoup4
detectlanguage
hikari-crescent
hikari-miru